fix: move SIGWINCH handling to platform-specific files for Windows build
