+++ internal/occ/cmd/component/exec.go
@@ -249,22 +249,7 @@ func sendResize(ws *wsWriter, width, height uint16) {
 	_ = ws.write(websocket.BinaryMessage, msg)
 }
 
-func watchResize(ctx context.Context, ws *wsWriter, fd int) {
-	ch := make(chan os.Signal, 1)
-	signal.Notify(ch, syscall.SIGWINCH)
-	defer signal.Stop(ch)
-
-	for {
-		select {
-		case <-ctx.Done():
-			return
-		case <-ch:
-			if w, h, err := term.GetSize(fd); err == nil {
-				sendResize(ws, safeUint16(w), safeUint16(h))
-			}
-		}
-	}
-}
+// watchResize is defined in exec_resize_unix.go and exec_resize_windows.go
 
 func safeUint16(v int) uint16 {
 	if v < 0 {
