+++ internal/occ/cmd/component/exec_resize_windows.go
@@ -1,0 +1,13 @@
+// Copyright 2026 The OpenChoreo Authors
+// SPDX-License-Identifier: Apache-2.0
+
+//go:build windows
+
+package component
+
+import (
+	"context"
+)
+
+// watchResize is a no-op on Windows — SIGWINCH is not available.
+func watchResize(_ context.Context, _ *wsWriter, _ int) {}
